fix(footer): fix server client differences
