--- src/components/molecules/ZopaFooter/styles/Footer.tsx
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const Footer = styled.footer`
-  background-color: ${({ theme }) => theme.footer.bgColor};
-`;
