--- src/components/molecules/ZopaFooter/styles/Heading.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-import Heading from '../../../atoms/Heading/Heading';
-
-export const FooterHeading = styled(Heading)`
-  text-transform: uppercase;
-`;
