--- src/components/molecules/ZopaFooter/styles/Icon.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const Icon = styled.span<{ variant: string }>`
-  display: inline-block;
-  width: 32px;
-  height: 32px;
-  background: url(${({ variant }) => variant}) no-repeat;
-`;
