+++ src/components/molecules/ZopaFooter/styles/SocialLink.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
+import { Icon } from '../ZopaFooter';
 import styled from 'styled-components';
 import Link, { LinkProps } from '../../../atoms/Link/Link';
-import { Icon } from './Icon';
 import { spacing } from '../../../../constants';
 
 interface SocialLink extends LinkProps {
