--- src/components/molecules/ZopaFooter/styles/index.ts
@@ -1,8 +1,0 @@
-export { Footer } from './Footer';
-export { FooterHeading as Heading } from './Heading';
-export { Icon } from './Icon';
-export { LegalBlock } from './LegalBlock';
-export { List } from './List';
-export { LogoBlock } from './LogoBlock';
-export { SocialBlock } from './SocialBlock';
-export { SocialLink } from './SocialLink';
